it works except storybook doesn't accept promises
