+++ package.json
@@ -48,7 +48,6 @@
     "electron-updater": "^4.1.2",
     "fs-extra": "^10.0.1",
     "globby": "^13.1.1",
-    "intl-messageformat": "^9.12.0",
     "lit-html": "^2.2.1",
     "punycode": "^2.1.1",
     "ShadowsocksConfig": "Jigsaw-Code/outline-shadowsocksconfig#v0.1.2",
