+++ src/www/ui_components/server_card/stories.ts
@@ -18,9 +18,9 @@
 import {html} from "lit-html";
 
 import {makeStorybookConfig} from "../../.storybook/make_storybook_config";
+import {controls as serverConnectionVizControls} from "./server_connection_viz";
 import {Localized, languageControl, makeLocalize} from "../../.storybook/make_localize";
 import {ServerCard, ServerConnectionState} from "./index";
-import {controls as serverConnectionVizControls} from "./server_connection_viz/stories";
 
 export const Example = async ({language, serverName, serverAddress, state, expanded}: Localized<ServerCard>) => {
   const localize = await makeLocalize(language);
